+++ Data/Models/Order.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+
+namespace Data.Models
+{
+    public class Order
+    {
+        public int OrderId { get; set; }
+        public string CompanyName { get; set; }
+        public string Description { get; set; }
+        public decimal OrderTotal { get; set; }
+        public List<OrderProduct> OrderProducts { get; set; }
+    }
+}
