+++ Data/Models/OrderProduct.cs
@@ -1,0 +1,11 @@
+﻿namespace Data.Models
+{
+    public class OrderProduct
+    {
+        public int OrderId { get; set; }
+        public int ProductId { get; set; }
+        public Product Product { get; set; }
+        public int Quantity { get; set; }
+        public decimal Price { get; set; }
+    }
+}
