+++ Data/Models/Product.cs
@@ -1,0 +1,8 @@
+﻿namespace Data.Models
+{
+    public class Product
+    {
+        public string Name { get; set; }
+        public decimal Price { get; set; }
+    }
+}
